more unit test for Scan function

+++ CheckoutKata.Core/Models/PricingRule.cs
@@ -3,7 +3,7 @@ namespace CheckoutKata.Core.Models
     public class PricingRule
     {
         public string SKU { get; private set; }
-        public int UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public bool HasSpecialPrice { get; private set; }
         public SpecialPrice? SpecialPrice { get; set; }
